requirements.txt
Flask
faiss-cpu  # or faiss-gpu if you are using a GPU
sentence-transformers
